requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.54.3
sounddevice==0.4.6
numpy==1.24.3
soundfile==0.12.1
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.0
